+++ src/main/java/fr/unice/polytech/kaleate/outils/InterfacePayement.java
@@ -1,5 +1,5 @@
 package fr.unice.polytech.kaleate.outils;
 
 public interface InterfacePayement {
-    boolean payer(float solde);
+    boolean payer(double solde);
 }
