+++ src/main/java/fr/unice/polytech/kaleate/outils/Monnayable.java
@@ -5,19 +5,19 @@ public interface Monnayable {
      * prix tout compris dedans
      * @return float
      */
-    public float getPrix();
+    public double getPrix();
 
     /**
      * prix sans reduction, avec tous les éléments supplémentaires
      * @return float
      */
-    public float getPrixSansReduction();
+    public double getPrixSansReduction();
 
     /**
      * prix de base, sans réduction ni élément supplémentaire
      * @return float
      */
-    public float getPrixBase();
+    public double getPrixBase();
 
 
 }
