+++ src/main/java/fr/unice/polytech/kaleate/outils/PayementExterne.java
@@ -2,7 +2,7 @@ package fr.unice.polytech.kaleate.outils;
 
 public class PayementExterne implements InterfacePayement {
     @Override
-    public boolean payer(float solde){
+    public boolean payer(double solde){
 
         return true;
     }
